Version 6.0.5
Added more optional fancy window/buffer navigation:

You can turn on the ability to use control and the arrow keys to move between windows.
You can turn on the ability to use <C-TAB> and <C-S-TAB> to open the next and previous (respectively) buffer in the current window.
You can turn on the ability to use <C-TAB> and <C-S-TAB> to switch windows (forward and backwards respectively.)

Fixed an issue with window sizing when we run out of buffers.