Version 6.0.8
Apparently some VIM builds are having a hard time with line continuation in scripts so the few that were here have been removed.


Updated debugging code so that debug output is put into a buffer which can then be written to disk or emailed to me when someone is having a major issue. Can also write directly to a file (VERY SLOWLY) on UNIX or Win32 (not 95 or 98 at the moment) or use VIM's echo function to display the output to the screen.

Several people have had issues when the hidden option is turned on. So I have put in several checks to make sure folks know this if they try to use MBE with the option set.