Version 6.2.1
If MBE is the only window (because of :bd for example) and there are still eligible buffers then one of them will be displayed.

The <Leader>mbe mapping now highlights the buffer from the current window.

The delete ('d') binding in the MBE window now restors the cursor position, which can help if you want to delete several buffers in a row that are not at the beginning of the buffer list.

Added a new key binding ('p') in the MBE window to switch to the previous window (last edit window)