Version 6.2.2
Changed the way the g:miniBufExplorerMoreThanOne
global is handled. You can set this to the number
of eligible buffers you want to be loaded before
the MBE window is loaded. Setting it to 0 causes
the MBE window to be opened even if there are no
buffers. Setting it to 4 causes the window to stay
closed until the 4th eligible buffer is loaded.

Added a MinHeight option. This is nice if you want
the MBE window to always take the same amount of
space. For example set MaxHeight and MinHeight to 2
and set MoreThanOne to 0 and you will always have
a 2 row (plus the ruler :) MBE window.

I now setlocal foldcomun=0 and nonumber in the MBE
window. This is for those of you that like to have
these options turned on locally. I'm assuming noone
outthere wants foldcolumns and line numbers in the
MBE window? :)

Fixed a bug where an empty MBE window was taking half
of the screen (partly why the MinHeight option was
added.)