Version 6.3.1
Added g:miniBufExplForceSyntaxEnable option for folks that are experiencing the vim bug where buffers show up without highlighting when another buffer has been closed, quit, wiped or deleted. This resolves all known issues with buffers loosing their highlighting. If you find a case that is not resolved, please report it :)

Folded source so that it's easier to navigate